Check if request user is the owner of the resource

+++ loan_api/loans/views.py
@@ -1,4 +1,4 @@
-from django.http import Http404
+from django.shortcuts import get_object_or_404
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -10,33 +10,31 @@ from loans.models import Loan
 class LoanView(APIView):
     permission_classes = [IsAuthenticated]
 
-    def get_loan_by_id(self, id):
-        try:
-            loan = Loan.objects.get(pk=id)
-            return loan
-        except Loan.DoesNotExist:
-            raise Http404
-
     def get(self, request, id):
-        loan = self.get_loan_by_id(id)
-        serializer = LoanSerializer(loan, context={"request": request})
+        loan = get_object_or_404(Loan, pk=id)
+        if request.user != loan.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
+        serializer = LoanSerializer(loan, context={"request": request})
         return Response(serializer.data)
 
     def patch(self, request, id):
-        loan = self.get_loan_by_id(id)
+        loan = get_object_or_404(Loan, pk=id)
+        if request.user != loan.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
         serializer = LoanSerializer(loan, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
             return Response(serializer.data)
-
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, id):
-        loan = self.get_loan_by_id(id)
-        loan.delete()
+        loan = get_object_or_404(Loan, pk=id)
+        if request.user != loan.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
+        loan.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
